test(update,models): move remaining tests into src modules
Move test_update_version.rs into src/update/mod.rs and
test_models_serialization.rs into src/models/analysis.rs (where
CodeAnalysisRecord / CodeAnalysisToolCalls / ExtensionType and the
other tested types are defined).

Completes the per-file test relocation — no more tests/test_*.rs
files remain at the top level of tests/, only the integration_tests
entry point plus its integrations/ submodule.
